requirements.txt
aiogram==3.15.0
openai>=1.0.0
faster-whisper==1.1.1
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.0
requests>=2.31.0
